[UPDATE] pisi: Add the evobuild script
The evobuild script is currently in its early testing phase, but is quite
simply a clean-chroot based builder for Evolve OS. While it certainly takes
inspiration from chroot builders for other distros, their code was not
looked at because lets be honest, creating tools like this for PiSi is
a total pain in the arse. i.e. clean room implementation

This script is currently very delicate, and should be invoked in an
empty directory in case anything goes wrong (and you can then clean up)
One nice feature of the system is the caching of .eopkg's and source
tarballs for quicker rebuilds and initialisations.